requirements.txt
Django >= 3.2.4, <3.3
djangorestframework >= 3.12.4, <3.13
djangorestframework-jwt == 1.11.0
